+++ lib/radar/integration/rails3.rb
@@ -1,3 +1,5 @@
+require "rails"
+
 module Radar
   module Integration
     # Allows drop-in integration with Rails 3 for Radar. This
@@ -6,12 +8,11 @@ module Radar
     # information to the exception (such as the rack environment,
     # request URL, etc.)
     class Rails3
-      def self.integrate!(app, opts=nil)
-        opts ||= {}
-        raise ArgumentError.new("Must specify a `:rails_app` to point to your Rails 3 application.") if !opts[:rails_app]
+      def self.integrate!(app)
+        raise ArgumentError.new("Rails integration requires a Rails application to be defined.") if !Rails.application
 
         # For now just use the Rack::Radar
-        opts[:rails_app].config.middleware.use "Rack::Radar", :application => app
+        Rails.application.config.middleware.use "Rack::Radar", :application => app
       end
     end
   end