+++ test/radar/integration/rails3_test.rb
@@ -1,4 +1,5 @@
 require 'test_helper'
+require 'rails'
 
 class Rails3IntegrationTest < Test::Unit::TestCase
   context "rails3 integration class" do
@@ -7,10 +8,22 @@ class Rails3IntegrationTest < Test::Unit::TestCase
       @app = Radar::Application.new(:app, false)
     end
 
-    should "raise an argument error if an application is not specified" do
-      assert_raises(ArgumentError) { @klass.integrate!(@app, :rails_app => nil) }
+    teardown do
+      # HACK I don't think rails was ever intended to work this way,
+      # but I'm able to clear out the application using this.
+      Rails.application = nil
     end
 
-    # TODO: Add rails to gemfile and test it?
+    should "raise an argument error if an application is not specified" do
+      assert_raises(ArgumentError) { @klass.integrate!(@app) }
+    end
+
+    should "integrate with a rails application" do
+      # NOTE This test used to actually initialize the rails application
+      # but it was creating a `log/` directory in the toplevel which was
+      # pretty annoying. So this suffices for now.
+      rails_app = Class.new(Rails::Application)
+      assert_nothing_raised { @klass.integrate!(@app) }
+    end
   end
 end