+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.3.0 (unreleased)
 
-  - Added `IoReporter` to log to any `IO` object.
+  - Rails 3 integration. See user guide for more information.
   - Rack integration. See user guide for more information. [GH-13]
+  - Added `IoReporter` to log to any `IO` object.
   - Refinements to `FileReporter` and sprinkled logger statements around.
   - Lightweight logging mechanism added so that you can verify Radar is doing
     its job. [GH-9]